functions/tip_worker/requirements.txt
google-genai>=1.0.0
supabase>=2.0.0
functions-framework>=3.0.0
